fix(adr-019): address Phase B code review findings
10 issues from code review:

1. Correct misleading 'idempotent' JSDoc on closeSession in types.ts —
   adapter is always called; busy/cancelled flags are always cleared.

2. Handle terminal-state (COMPLETED/FAILED) descriptors in openSession:
   bypass SESSION_TRANSITIONS (which forbids terminal→RUNNING) via direct
   Map mutation + _persistDescriptor, matching the closeStory pattern.

3. Add comment to mock-session-manager runInSession stub noting the
   AgentResult vs TurnResult shape mismatch for Phase B test authors.

4. Export NO_OP_INTERACTION_HANDLER + InteractionHandler type from the
   agents barrel (src/agents/index.ts) so callers don't import internals.

5. Fix barrel import violation in manager.ts: import NO_OP_INTERACTION_HANDLER
   from '../agents' not '../agents/interaction-handler'.

6. Add clarifying comment in adapter.ts openSession that opts.resume is
   advisory — ACP is inherently self-resuming via ensureAcpSession.

7. Add Phase B types to src/session/index.ts barrel: OpenSessionRequest,
   SendPromptOpts, RunInSessionOpts, NameForRequest.

8. Extract _runInSessionLegacy to src/session/manager-legacy.ts reducing
   manager.ts from 729→613 lines. Split manager-phase-b.test.ts (491 lines)
   into manager-phase-b-session.test.ts and manager-phase-b-prompt.test.ts.

9. Add test for adapter.openSession throw path (descriptor must not be RUNNING).
   Add test for terminal-state re-open (descriptor resets to RUNNING).

10. Add test for maxTurns passthrough to adapter.sendTurn.
