+++ src/agents/index.ts
@@ -1,4 +1,6 @@
 export type { AgentAdapter, AgentCapabilities, AgentResult, AgentRunOptions, CompleteOptions } from "./types";
+export type { InteractionHandler } from "./interaction-handler";
+export { NO_OP_INTERACTION_HANDLER } from "./interaction-handler";
 export { CompleteError } from "./types";
 export { getAllAgentNames, getInstalledAgents, checkAgentHealth, KNOWN_AGENT_NAMES } from "./registry";
 export type { ModelCostRates, TokenUsage, CostEstimate, TokenUsageWithConfidence, SessionTokenUsage } from "./cost";
