+++ src/session/manager.ts
@@ -14,13 +14,14 @@
 import { createHash, randomUUID } from "node:crypto";
 import { mkdir } from "node:fs/promises";
 import { isAbsolute, join, relative, sep } from "node:path";
-import { NO_OP_INTERACTION_HANDLER } from "../agents/interaction-handler";
+import { NO_OP_INTERACTION_HANDLER } from "../agents";
 import type { AgentRunRequest, IAgentManager } from "../agents/manager-types";
 import type { AgentAdapter, AgentResult, SessionHandle, TurnResult } from "../agents/types";
 import type { NaxConfig } from "../config";
 import { resolvePermissions } from "../config/permissions";
 import { NaxError } from "../errors";
 import { getLogger } from "../logger";
+import { runInSessionLegacy } from "./manager-legacy";
 import type {
   CreateSessionOptions,
   ISessionManager,
@@ -433,6 +434,16 @@ export class SessionManager implements ISessionManager {
       this.transition(created.id, "RUNNING");
     } else if (existingDescriptor.state === "CREATED") {
       this.transition(existingDescriptor.id, "RUNNING");
+    } else if (existingDescriptor.state === "COMPLETED" || existingDescriptor.state === "FAILED") {
+      // Terminal → RUNNING is not a valid state-machine transition, so bypass
+      // transition() and update directly (same pattern as closeStory).
+      const updated: SessionDescriptor = {
+        ...existingDescriptor,
+        state: "RUNNING",
+        lastActivityAt: _sessionManagerDeps.now(),
+      };
+      this._sessions.set(existingDescriptor.id, updated);
+      this._persistDescriptor(updated);
     }
 
     getLogger().debug("session", "Session opened via SessionManager", {
@@ -564,136 +575,26 @@ export class SessionManager implements ISessionManager {
     }
   }
 
-  // ─── Legacy runInSession body ───────────────────────────────────────────────
-
-  /**
-   * Per-session lifecycle primitive — see ISessionManager for full contract.
-   *
-   * Bookkeeping order:
-   *   1. CREATED → RUNNING (only if currently CREATED; RESUMING is left alone
-   *      so rectification loops that re-enter an already-RUNNING session don't
-   *      throw SESSION_INVALID_TRANSITION).
-   *   2. Call the runner. Result.protocolIds (if present) is bound via
-   *      bindHandle so the disk descriptor captures the audit correlation.
-   *   3. RUNNING → COMPLETED on success, RUNNING → FAILED on failure. If the
-   *      runner threw, we transition to FAILED and re-throw.
-   *
-   * ADR-013 Phase 1: accepts IAgentManager + AgentRunRequest instead of the
-   * raw SessionAgentRunner function. Calls agentManager.run(injectedRequest)
-   * after injecting the onSessionEstablished callback for eager handle binding.
-   */
   private async _runInSessionLegacy(
     id: string,
     agentManager: IAgentManager,
     request: AgentRunRequest,
     _options?: SessionRunOptions,
   ): Promise<AgentResult> {
-    const pre = this._sessions.get(id);
-    if (!pre) {
-      throw new NaxError(`Session "${id}" not found in registry`, "SESSION_NOT_FOUND", {
-        stage: "session",
-        sessionId: id,
-      });
-    }
-
-    if (pre.state === "CREATED") {
-      this.transition(id, "RUNNING");
-    }
-
-    // #591: inject onSessionEstablished so the adapter can bind protocolIds
-    // eagerly — before any prompt runs. If the run is interrupted between
-    // session-established and result-returned, the descriptor already
-    // carries the correlation needed to resume. The caller's own callback
-    // (if any) is chained afterwards so both fire.
-    const callerCallback = request.runOptions.onSessionEstablished;
-    const injectedRequest: AgentRunRequest = {
-      ...request,
-      runOptions: {
-        ...request.runOptions,
-        onSessionEstablished: (protocolIds, sessionName) => {
-          try {
-            this.bindHandle(id, sessionName, protocolIds);
-          } catch (err) {
-            getLogger().warn("session", "bindHandle via onSessionEstablished failed", {
-              storyId: this._sessions.get(id)?.storyId,
-              sessionId: id,
-              error: err instanceof Error ? err.message : String(err),
-            });
-          }
-          callerCallback?.(protocolIds, sessionName);
-        },
+    return runInSessionLegacy(
+      id,
+      agentManager,
+      request,
+      {
+        sessions: this._sessions,
+        transition: this.transition.bind(this),
+        bindHandle: this.bindHandle.bind(this),
+        handoff: this.handoff.bind(this),
+        persistDescriptor: this._persistDescriptor.bind(this),
+        now: _sessionManagerDeps.now,
       },
-    };
-
-    const config = request.runOptions.config;
-    const maxRetriable = config?.execution?.sessionErrorRetryableMaxRetries ?? 3;
-    const maxNonRetriable = config?.execution?.sessionErrorMaxRetries ?? 1;
-    let sessionRetries = 0;
-
-    let result: AgentResult;
-    // Session-transport retry loop: retries only on fail-adapter-error.
-    // Auth/rate-limit failures surface immediately so AgentManager's fallback
-    // and backoff logic fires without a SessionManager-level retry doubling up.
-    while (true) {
-      try {
-        result = await agentManager.run(injectedRequest);
-      } catch (err) {
-        if (this._sessions.get(id)?.state === "RUNNING") {
-          this.transition(id, "FAILED");
-        }
-        throw err;
-      }
-
-      if (result.adapterFailure?.outcome === "fail-adapter-error") {
-        const max = result.adapterFailure.retriable ? maxRetriable : maxNonRetriable;
-        if (sessionRetries < max && !request.signal?.aborted) {
-          sessionRetries++;
-          getLogger().warn("session", "Session transport error — retrying with fresh session", {
-            sessionId: id,
-            storyId: this._sessions.get(id)?.storyId,
-            retriable: result.adapterFailure.retriable,
-            attempt: sessionRetries,
-            maxAttempts: max,
-          });
-          continue;
-        }
-      }
-      break;
-    }
-
-    if (result.protocolIds) {
-      const current = this._sessions.get(id);
-      const handle = current?.handle;
-      if (handle) {
-        this.bindHandle(id, handle, result.protocolIds);
-      } else {
-        const updated: SessionDescriptor = {
-          ...(current as SessionDescriptor),
-          protocolIds: result.protocolIds,
-          lastActivityAt: _sessionManagerDeps.now(),
-        };
-        this._sessions.set(id, updated);
-        this._persistDescriptor(updated);
-      }
-    }
-
-    // Gap A: reconcile descriptor.agent after an agent swap in AgentManager.
-    // agentFallbacks.at(-1).newAgent is the final agent used; handoff() updates
-    // the descriptor so crash recovery and metrics see the correct agent name.
-    const finalAgent = result.agentFallbacks?.at(-1)?.newAgent;
-    if (finalAgent) {
-      const current = this._sessions.get(id);
-      if (current && finalAgent !== current.agent) {
-        this.handoff(id, finalAgent, "post-run-reconcile");
-      }
-    }
-
-    const current = this._sessions.get(id);
-    if (current?.state === "RUNNING") {
-      this.transition(id, result.success ? "COMPLETED" : "FAILED");
-    }
-
-    return result;
+      _options,
+    );
   }
 
   sweepOrphans(ttlMs = DEFAULT_ORPHAN_TTL_MS): number {
